Некоторые примечания
