+++ app/src/main/java/com/example/popularlibraries/presenter/DetailPresenter.kt
@@ -1,12 +1,12 @@
 package com.example.popularlibraries.presenter
 
 import com.example.popularlibraries.view.IScreens
-import com.example.popularlibraries.view.main.MainView
+import com.example.popularlibraries.view.details.DetailsView
 import com.github.terrakok.cicerone.Router
 import moxy.MvpPresenter
 
-class MainPresenter(val router: Router, val screens: IScreens) :
-    MvpPresenter<MainView>() {
+class DetailPresenter(val router: Router, val screens: IScreens) :
+    MvpPresenter<DetailsView>() {
 
     override fun onFirstViewAttach() {
         super.onFirstViewAttach()
