+++ app/src/main/java/com/example/popularlibraries/view/details/DetailsView.kt
@@ -1,0 +1,13 @@
+package com.example.popularlibraries.view.details
+
+import moxy.MvpView
+import moxy.viewstate.strategy.AddToEndSingleStrategy
+import moxy.viewstate.strategy.StateStrategyType
+
+//Так как всё, что появится на экране — просто логин, интерфейс включает всего 1 метод:
+//● init() — для первичной инициализации списка, который мы будем вызывать при
+//присоединении View к Presenter;
+@StateStrategyType(AddToEndSingleStrategy::class)
+interface DetailsView : MvpView {
+    fun init()
+}
