+++ app/src/main/java/com/example/popularlibraries/view/users/UsersRVAdapter.kt
@@ -1,12 +1,17 @@
-package com.example.popularlibraries.view
+package com.example.popularlibraries.view.users
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.example.popularlibraries.databinding.ItemUserBinding
+import com.example.popularlibraries.presenter.IItemView
 import com.example.popularlibraries.presenter.IUserListPresenter
 
 //Адаптер не имеет ссылок на данные и полностью делегирует процесс наполнения View в Presenter.
+//Адаптер — это фреймворковая вещь, а, значит, он —
+//ui. Соответственно, логики и данных прямо в нём быть не должно.
+//Может показаться, что если адаптер — ui, то он и
+//есть View. Но это не так.
 
 class UsersRVAdapter(val presenter: IUserListPresenter) :
     RecyclerView.Adapter<UsersRVAdapter.ViewHolder>() {
@@ -19,6 +24,12 @@ class UsersRVAdapter(val presenter: IUserListPresenter) :
         override fun setLogin(text: String) = with(binding) {
             loginTextView.text = text
         }
+    }
+
+    //ViewHolder имплементит наш интерфейс UserItemView с методом setLogin.
+    //А UserItemView имплементит IItemView с позицией элемента списка var pos.
+    interface UserItemView : IItemView {
+        fun setLogin(text: String)
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
