+++ app/src/main/java/com/example/popularlibraries/view/users/UsersView.kt
@@ -1,4 +1,4 @@
-package com.example.popularlibraries.view
+package com.example.popularlibraries.view.users
 
 import moxy.MvpView
 import moxy.viewstate.strategy.AddToEndSingleStrategy
@@ -8,6 +8,7 @@ import moxy.viewstate.strategy.StateStrategyType
 //● init() — для первичной инициализации списка, который мы будем вызывать при
 //присоединении View к Presenter;
 //● updateList() — для обновления содержимого списка.
+//Эти методы будет выполнять наша viewState
 @StateStrategyType(AddToEndSingleStrategy::class)
 interface UsersView : MvpView {
     fun init()
