+++ app/src/main/res/navigation/nav_graph.xml
@@ -6,7 +6,7 @@
 
     <fragment
         android:id="@+id/usersFragment"
-        android:name="com.example.popularlibraries.view.UsersFragment"
+        android:name="com.example.popularlibraries.view.users.UsersFragment"
         android:label="@string/app_name"
         tools:layout="@layout/fragment_users">
         <action
@@ -15,7 +15,7 @@
     </fragment>
     <fragment
         android:id="@+id/detailFragment"
-        android:name="com.example.popularlibraries.view.DetailFragment"
+        android:name="com.example.popularlibraries.view.details.DetailFragment"
         android:label="Подробнее"
         tools:layout="@layout/fragment_details">
         <argument
